+++ collector-exoplayer-example/src/main/java/com/bitmovin/analytics/exoplayer/example/MainActivity.kt
@@ -186,7 +186,6 @@ class MainActivity : AppCompatActivity(), DebugListener, Player.Listener {
         bitmovinAnalyticsConfig.customData6 = "customData6"
         bitmovinAnalyticsConfig.customData7 = "customData7"
         bitmovinAnalyticsConfig.path = "/vod/new/"
-        bitmovinAnalyticsConfig.heartbeatInterval = 59700
         bitmovinAnalyticsConfig.ads = false
         bitmovinAnalyticsConfig.isLive = false